moving common methods to drutils file
